+++ algebraeon/src/integer_polynomial.rs
@@ -117,6 +117,10 @@ impl PythonStructure for PythonIntegerPolynomial {
     fn inner(&self) -> &<Self::Structure as SetSignature>::Set {
         &self.inner
     }
+
+    fn into_inner(self) -> <Self::Structure as SetSignature>::Set {
+        self.inner
+    }
 }
 
 impl_pymethods_elem!(PythonIntegerPolynomial);