+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project should be documented in this file.
 
 ## [Unreleased]
 
+ - Allow explicitly converting from a larger set to a smaller set.
+
 ## [0.0.2] - 2026-01-15
 
  - Added polynomials over `Rat`.