+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project should be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.3] - 2026-03-06
+
  - Allow explicitly converting from a larger set to a smaller set.
  - Added `RealAlg` for real algebraic numbers.
  - Added `CpxAlg` for real algebraic numbers.